test, commenting the second way of sending email for now because princess still sending

+++ app/Console/Commands/SendAgencies.php
@@ -41,64 +41,72 @@ class SendAgencies extends Command
      */
     public function handle()
     {
-        $applicant = Applicant::where('status', 'open')->get()->first();
-        
-        $exists = Storage::disk('local')->exists('sendagencies.txt');
-        
-        if(!$applicant && $exists) {
+        // $use set to false because we dont want to use it for now because princess still sending email
+        // use this is you want to use the second  way of sending email, second way is to send email daily or scheduled
+        // remove $use if you want to use second way
+        $use = false;
+        if($use) {
+
+
+            $applicant = Applicant::where('status', 'open')->get()->first();
             
-            $contents = Storage::disk('local')->get('sendagencies.txt');
-            Storage::disk('local')->delete('sendagencies.txt');  
-            Storage::disk('local')->put('sendagencies_done.txt', $contents);
-        }
-        
-        ini_set('max_execution_time', 3600); // 1 hr
-
-        if($applicant) {
-
-            $last_emailed_agency = $applicant->last_emailed_agency;
-
-            if(!$last_emailed_agency) {
-                $last_emailed_agency = 0;
+            $exists = Storage::disk('local')->exists('sendagencies.txt');
+            
+            if(!$applicant && $exists) {
+                
+                $contents = Storage::disk('local')->get('sendagencies.txt');
+                Storage::disk('local')->delete('sendagencies.txt');  
+                Storage::disk('local')->put('sendagencies_done.txt', $contents);
             }
-
-            $agencies = Agency::where('id', '>', $last_emailed_agency)->paginate(200);
-
-            $last_agency_id = 0;
-
-            foreach ($agencies as $agency) {
-
-                $emails = json_decode($agency->email);
             
-                foreach ($emails as $v) {
+            ini_set('max_execution_time', 3600); // 1 hr
 
-                    //Mail::to($v)->send(new PrincessSendAgiences($this->data));
+            if($applicant) {
+
+                $last_emailed_agency = $applicant->last_emailed_agency;
+
+                if(!$last_emailed_agency) {
+                    $last_emailed_agency = 0;
                 }
 
-                $time = date("h:i:sa").' via build ';
-                $exists = Storage::disk('local')->exists('sendagencies.txt');
-                if(!$exists) {
-                    Storage::disk('local')->put('sendagencies.txt', "\n".$time.' '.$agency->email.' ['.$agency->id.']');
+                $agencies = Agency::where('id', '>', $last_emailed_agency)->paginate(200);
+
+                $last_agency_id = 0;
+
+                foreach ($agencies as $agency) {
+
+                    $emails = json_decode($agency->email);
+                
+                    foreach ($emails as $v) {
+
+                        //Mail::to($v)->send(new PrincessSendAgiences($this->data));
+                    }
+
+                    $time = date("h:i:sa").' via build ';
+                    $exists = Storage::disk('local')->exists('sendagencies.txt');
+                    if(!$exists) {
+                        Storage::disk('local')->put('sendagencies.txt', "\n".$time.' '.$agency->email.' ['.$agency->id.']');
+                    } else {
+                        $contents = Storage::disk('local')->get('sendagencies.txt');
+                        Storage::disk('local')->put('sendagencies.txt', $contents."\n".$time.' '.$agency->email.' ['.$agency->id.']');
+                    }
+
+                    $last_agency_id = $agency->id;
+                }
+
+                $last_agency = Agency::orderBy('id', 'desc')->first();
+                
+                $applicant->send_end = date("Y-m-d").' '.date("H:i:s");
+
+                if($last_agency->id == $last_agency_id) {
+                    $applicant->status = 'close';
+                    $applicant->last_emailed_agency = 0;
                 } else {
-                    $contents = Storage::disk('local')->get('sendagencies.txt');
-                    Storage::disk('local')->put('sendagencies.txt', $contents."\n".$time.' '.$agency->email.' ['.$agency->id.']');
+                    $applicant->last_emailed_agency = $last_agency_id;  
                 }
-
-                $last_agency_id = $agency->id;
+                
+                $applicant->save();
             }
-
-            $last_agency = Agency::orderBy('id', 'desc')->first();
-            
-            $applicant->send_end = date("Y-m-d").' '.date("H:i:s");
-
-            if($last_agency->id == $last_agency_id) {
-                $applicant->status = 'close';
-                $applicant->last_emailed_agency = 0;
-            } else {
-                $applicant->last_emailed_agency = $last_agency_id;  
-            }
-            
-            $applicant->save();
         }
     }
 }
